Refactor AmbientSounds component
