--- frontend/components/FocusSpace/AmbientSounds/index.js
@@ -1,2 +1,0 @@
-export * from './AmbientSounds'
-export { default } from './AmbientSounds'
