+++ frontend/components/FocusSpace/AmbientSoundsPlayer/PlayButton.js
