+++ frontend/components/FocusSpace/AmbientSoundsPlayer/SoundWaveIcon.js
