+++ frontend/components/FocusSpace/AmbientSoundsPlayer/index.js
@@ -1,0 +1,2 @@
+export * from './AmbientSoundsPlayer'
+export { default } from './AmbientSoundsPlayer'
